+++ src/main/java/com/secondcommit/forum/controllers/PostController.java
@@ -1,6 +1,7 @@
 package com.secondcommit.forum.controllers;
 
 import com.secondcommit.forum.dto.PostDto;
+import com.secondcommit.forum.repositories.PostRepository;
 import com.secondcommit.forum.security.payload.MessageResponse;
 import com.secondcommit.forum.services.post.PostServiceImpl;
 import io.swagger.annotations.ApiOperation;
@@ -17,9 +18,11 @@ import org.springframework.web.bind.annotation.*;
 public class PostController {
 
     private final PostServiceImpl postService;
+    private final PostRepository postRepository;
 
-    public PostController(PostServiceImpl postService) {
+    public PostController(PostServiceImpl postService, PostRepository postRepository) {
         this.postService = postService;
+        this.postRepository = postRepository;
     }
 
     /**
@@ -30,7 +33,7 @@ public class PostController {
      */
     @PreAuthorize("hasAuthority('USER')")
     @PostMapping("/")
-    @ApiOperation("Creates new post")
+    @ApiOperation("Creates new post. Authentication required (USER)")
     public ResponseEntity<?> newPost(@RequestBody PostDto postDto, @CurrentSecurityContext(expression="authentication?.name") String username) {
 
         //Validates post
@@ -47,20 +50,36 @@ public class PostController {
      */
     @PreAuthorize("hasAuthority('USER')")
     @GetMapping("/{id}")
-    @ApiOperation("Gets POST")
+    @ApiOperation("Gets post. Authentication required (USER)")
     public ResponseEntity<?> getPost(@PathVariable Long id){
+
+        //Validates post
+        if (!postRepository.existsById(id))
+            return ResponseEntity.badRequest().body(new MessageResponse("Wrong id"));
+
         return postService.getPost(id);
     }
 
-    @PreAuthorize("hasAuthority('ADMIN')")
+    /**
+     * Method to update post.
+     * @param id
+     * @param postDto
+     * @param username (Gets from the jwt token)
+     * @return ResponseEntity (ok: post, bad request: messageResponse)
+     */
+    @PreAuthorize("hasAuthority('USER')")
     @PutMapping("/{id}")
-    @ApiOperation("Updates the post")
+    @ApiOperation("Updates the post. Authentication required (USER)")
     public ResponseEntity<?> updatePost(@PathVariable Long id, @RequestBody PostDto postDto,
                                         @CurrentSecurityContext(expression="authentication?.name") String username){
 
         //Validates Dto
         if (postDto.getTitle() == null && postDto.getContent() == null)
             return ResponseEntity.badRequest().body(new MessageResponse("Missing"));
+
+        //Validates id
+        if (!postRepository.existsById(id))
+            return ResponseEntity.badRequest().body(new MessageResponse("Wrong id"));
 
         return postService.updatePost(id, postDto, username);
     }
@@ -69,36 +88,73 @@ public class PostController {
      * Method to delete post
      * @param id
      * @param username (Gets from the jwt token)
-     * @return ResponseEntity
+     * @return ResponseEntity (messageResponse)
      */
     @PreAuthorize("hasAuthority('USER')")
     @DeleteMapping("/{id}")
-    @ApiOperation("Deletes the post")
+    @ApiOperation("Deletes the post. Authentication required (USER)")
     public ResponseEntity<?> deletePost(Long id,
                                         @CurrentSecurityContext(expression="authentication?.name") String username){
+
+        //Validates id
+        if (!postRepository.existsById(id))
+            return ResponseEntity.badRequest().body(new MessageResponse("Wrong id"));
+
         return postService.deletePost(id, username);
     }
 
+    /**
+     * Method to add or remove like from post
+     * @param id
+     * @param username (Gets from the jwt token)
+     * @return ResponseEntity(ok: totalLikes, bad request: messageResponse)
+     */
     @PreAuthorize("hasAuthority('USER')")
     @PutMapping("/like/{id}")
-    @ApiOperation("Adds or removes like, depending on the previous state")
+    @ApiOperation("Adds or removes like, depending on the previous state. Authentication required (USER)")
     public ResponseEntity<?> like(Long id,
                                         @CurrentSecurityContext(expression="authentication?.name") String username){
+
+        //Validates id
+        if (!postRepository.existsById(id))
+            return ResponseEntity.badRequest().body(new MessageResponse("Wrong id"));
+
         return postService.like(id, username);
     }
 
+    /**
+     * Method to add or remove dislike from post
+     * @param id
+     * @param username (Gets from the jwt token)
+     * @return ResponseEntity(ok: totalDislikes, bad request: messageResponse)
+     */
     @PreAuthorize("hasAuthority('USER')")
     @PutMapping("/dislike/{id}")
-    @ApiOperation("Adds or removes dislike, depending on the previous state")
+    @ApiOperation("Adds or removes dislike, depending on the previous state. Authentication required (USER)")
     public ResponseEntity<?> dislike(Long id,
                                   @CurrentSecurityContext(expression="authentication?.name") String username){
+
+        //Validates id
+        if (!postRepository.existsById(id))
+            return ResponseEntity.badRequest().body(new MessageResponse("Wrong id"));
+
         return postService.dislike(id, username);
     }
 
+    /**
+     * Method to fix or unfix a post. ADMIN only
+     * @param id
+     * @return ResponseEntity (MessageResponse)
+     */
     @PreAuthorize("hasAuthority('ADMIN')")
     @PutMapping("/fix/{id}")
-    @ApiOperation("Fix post. Only admins are allowed")
+    @ApiOperation("Fix post. . Authentication required (ADMIN)")
     public ResponseEntity<?> fix(Long id){
+
+        //Validates id
+        if (!postRepository.existsById(id))
+            return ResponseEntity.badRequest().body(new MessageResponse("Wrong id"));
+
         return postService.fix(id);
     }
 }
